added working home screen
