+++ src/containers/HomeScreen.js
@@ -1,0 +1,48 @@
+import React, { Component } from 'react';
+import { View, Text, StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'center',
+        backgroundColor: '#F5FCFF',
+    },
+    welcome: {
+        fontSize: 20,
+        textAlign: 'center',
+        margin: 10,
+    },
+    instructions: {
+        textAlign: 'center',
+        color: '#333333',
+        marginBottom: 5,
+    },
+});
+
+class HomeScreen extends React.Component {
+  static route = {
+    navigationBar: {
+      title: 'Home',
+    }
+  }
+
+  render() {
+    return (
+      <View style={styles.container}>
+        <Text style={styles.welcome}>
+            Welcome to React Native!
+        </Text>
+        <Text style={styles.instructions}>
+            To get started, edit index.ios.js
+        </Text>
+        <Text style={styles.instructions}>
+            Press Cmd+R to reload,{'\n'}
+            Cmd+D or shake for dev menu
+        </Text>
+      </View>
+    );
+  }
+}
+
+export default HomeScreen;
