+++ src/index.js
@@ -11,47 +11,28 @@ import {
     Text,
     View
 } from 'react-native';
-import TabScreen from './components/TabScreen';
+import {
+  NavigationProvider,
+  StackNavigation,
+  TabNavigation,
+  TabNavigationItem as TabItem,
+} from '@exponent/ex-navigation';
+
+// import TabScreen from './components/TabScreen';
+import Router from './router';
+
 
 export default class hackathonApp extends Component {
-    render() {
-        console.log('objectasdfsad');
-        return (
-            <View style={styles.container}>
-            <TabScreen>
-                <Text style={styles.welcome}>
-                    Welcome to React Native!
-                </Text>
-                <Text style={styles.instructions}>
-                    To get started, edit index.ios.js
-                </Text>
-                <Text style={styles.instructions}>
-                    Press Cmd+R to reload,{'\n'}
-                    Cmd+D or shake for dev menu
-                </Text>
-            </TabScreen>
-            </View>
-        );
-    }
+  render() {
+    console.log('objectasdfsad');
+    return (
+      <NavigationProvider router={Router}>
+        <StackNavigation initialRoute={Router.getRoute('home')} />
+      </NavigationProvider>
+    );
+  }
 }
 
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-        backgroundColor: '#F5FCFF',
-    },
-    welcome: {
-        fontSize: 20,
-        textAlign: 'center',
-        margin: 10,
-    },
-    instructions: {
-        textAlign: 'center',
-        color: '#333333',
-        marginBottom: 5,
-    },
-});
+
 
 AppRegistry.registerComponent('hackathonApp', () => hackathonApp);
